build: Updating dependencies
composer:
* mediawiki/mediawiki-phan-config: 0.8.0 → 0.9.0
* mediawiki/minus-x: 0.3.1 → 0.3.2
* mediawiki/mediawiki-codesniffer: 28.0.0 → 29.0.0
* jakub-onderka/php-console-highlighter: 0.3.2 → 0.4.0

npm:
* grunt-banana-checker: 0.5.0 → 0.8.1

+++ includes/TheWikipediaLibraryHooks.php
@@ -89,8 +89,8 @@ class TheWikipediaLibraryHooks {
 				if ( !$globalUser->isAttached() ) {
 					return;
 				}
-				$accountAge = wfTimestamp( TS_UNIX ) -
-					wfTimestamp( TS_UNIX, $globalUser->getRegistration() );
+				$accountAge = (int)wfTimestamp( TS_UNIX ) -
+					(int)wfTimestamp( TS_UNIX, $globalUser->getRegistration() );
 				$minimumAge = $wgTwlRegistrationDays * 24 * 3600;
 				if ( $globalUser->getGlobalEditCount() >= $wgTwlEditCount && $accountAge >= $minimumAge ) {
 					EchoEvent::create( [
